Add DB storage for challenges and solutions.
When requesting a solution for a Pyramid Solitaire game, the service
will first check if the solution already exists in the database and
return the solution immediately if it does.  Otherwise it will
calculate the solution and store it for future requests before returning
the solution.

This uses PostgreSQL but a programmer should be able to change it
quickly if they want.