Return "Lose Quickly" when there is no solution.
Board challenges are the types where it may be impossible to have a
solution at all.  Instead of returning an empty list of solutions which
may be confusing, return a solution with one step called "Lose Quickly"
and a description saying there is no solution to make it clear there
isn't a solution.