Add REST behavior for long running jobs.
Based on https://farazdagi.com/2014/rest-and-long-running-jobs/, you can
GET a challenge and either get a 200 with the solution or 404.  You can
POST a challenge and get a 200 with the solution if it exists already,
or a 202 (Accepted) with a Location to check on its status.
Then you can GET the status URI and see a 200 with status pending, or a
303 (See Other) and a link to get the solution.